[Warp Specialization] Split partitioning into a separate pass (NFC)
